OneDrive - Greenstep Oy/repositories/Monitor/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable>=1.0.0b7pip
azure.identity
azure-keyvault-secrets
pyodbc
regex
datetime
requests
pandas
azure-storage-blob==2.1.0